Inline Copilot adapter implementation
