+++ cli/src/adapters/copilot-local/format-event.ts
