--- packages/adapters/copilot-local/src/cli/index.ts
@@ -1,1 +1,0 @@
-export { printCopilotStreamEvent } from "./format-event.js";
