--- packages/adapters/copilot-local/src/ui/index.ts
@@ -1,2 +1,0 @@
-export { buildCopilotLocalConfig } from "./build-config.js";
-export { createCopilotStdoutParser, parseCopilotStdoutLine } from "./parse-stdout.js";
