+++ ui/src/adapters/copilot-local/index.ts
@@ -1,9 +1,6 @@
 import type { UIAdapterModule } from "../types";
-import {
-  buildCopilotLocalConfig,
-  createCopilotStdoutParser,
-  parseCopilotStdoutLine,
-} from "@paperclipai/adapter-copilot-local/ui";
+import { buildCopilotLocalConfig } from "./build-config";
+import { createCopilotStdoutParser, parseCopilotStdoutLine } from "./parse-stdout";
 import { CopilotLocalConfigFields } from "./config-fields";
 
 export const copilotLocalUIAdapter: UIAdapterModule = {
