+++ ui/src/adapters/copilot-local/parse-stdout.ts
