+++ vitest.config.ts
@@ -4,7 +4,6 @@ export default defineConfig({
   test: {
     projects: [
       "packages/db",
-      "packages/adapters/copilot-local",
       "packages/adapters/codex-local",
       "packages/adapters/opencode-local",
       "server",
